+++ spy/tests/compiler/operator/test_attrop.py
