+++ spy/tests/compiler/operator/test_callop.py
